Add files via upload
description: There is one small ball, and one big ball. Small ball runs manually by the user, and the big ball is run by the computer. There is a option for recording the score, and lastly, a powerup  ball, which, upon clashed by the small ball, eradicates all the big balls except one.